Make the compiler return exit code 1 on any compilation error
The compiler used to return 1 and 2 for, respectively, syntax and semantic
errors, and the test suite used these to check if the test programs
were rejected where they should. While a bit cleaner, it made the
compiler somewhat coupled to the test code. Now the test suite rely on
System.err output instead.