Add class that models a JVM method
Code for a function is generated by injecting a JvmMethod instance
target into the statement and expression generators.
This is done to avoid global state and makes it easier to separate the
visitor classes into StatementGenerator and ExpressionGenerator.

The symbol table class that resolved functions to JVM method identifiers
and variables to addresses is discarded. The lookup table for functions
is put in the expression generator class, because it is only needed when
generating function calls. The lookup table for variables becomes part
of the JVM method class.