Handle int to double widening conversions
In an expression such as

    1 op 2.0

or
    expectsDouble(1)

a widening conversion needs to be performed on the int-valued expression
in order for the generated bytecode instructions for doubles to work.
This is done with the instruction i2d, which is added during
code generation right after the instruction that pushes the
subexpression onto the stack. If a conversion should occur, the code
generator is made aware of it from a boolean flag in the AST node of the
subexpression, which has been set during semantic analysis.