Use static type annotations of functions more effectively
Replace: notc.semantics.FunctionType -> notc.antlrgen.NotCParser.Signature

A function's signature only depends on its declaration, so instantiation
of the convenience class for signatures can be done by the parser.

Class Signature is defined alongside Type and they serve a similar purpose:
Replacing ANTLR-generated AST properties for typing with code that's
easier for the compiler components to work with.