Resolve function references to function headers
When lookupFunction(id) is called on a symbol table, it now returns the
function's original header, not its signature.

Class SymbolTable becomes more cohesive as it now only operates on
declarations as they were parsed, not a mix of that and a custom
convenience class (i.e. Signature, which is removed).

This breaks out the function header part of the function definition
parser rule. The new parser rule is used to instantiate the headers
of the built-in functions by parsing a resource file containing them.
This avoids having to create them in the symbol table's constructor.