Put operation constants in an enum
A String member stores the mnemonic, so now a lot of literals are not
repeated everywhere.

The enum stores the impact each operation has on the stack depth,
which is constant (except for invocations).  This avoids having to pass
the stack change as int literals every time a new instruction is added.

Also, add some lookup tables for retrieving operations associated
with different types. This gets rid of the type prefix previously stored
in the Type enum (e.g. i for int that was read when generating e.g.
"istore").