Allow any numerical expression to be negated
Previously, only number literals could be negative because they
were matched in a lexer rule. This caused an issue when the lexer
scanned an expression like

    1-1 // No whitespace inbetween

This was tokenized as

    '1', '-1'

and not

    '1', '-', '1'

since '-1' is the longer token. To avoid this, the negation operator
is now matched in a parsing rule for expressions. This allows any
expression to be parsed as negated, and attempts to negate non-numerical
ones are caught during semantic analysis.