Revert to Jasmin for bytecode assembly
Because it has a better public API than AsmTools.

Jasmin requires field references to be qualified with the name
of the class. This makes it trickier to have the built-in reading
methods share a common BufferedReader instance as the name of the
generated class is not known in advance. These new implementations
use local System.in Scanners instead and simply don't close them.